Rename maya-volume-exporter to maya-exporter
Changes to be committed:
	modified:   ../GNUmakefile
	modified:   ../buildscripts/exporter/Dockerfile
	modified:   ../buildscripts/exporter/build.sh
	renamed:    maya-volume-exporter/app/collector/collector.go -> maya-exporter/app/collector/collector.go
	renamed:    maya-volume-exporter/app/collector/collector_test.go -> maya-exporter/app/collector/collector_test.go
	renamed:    maya-volume-exporter/app/command/command.go -> maya-exporter/app/command/command.go
	renamed:    maya-volume-exporter/app/command/exporter.go -> maya-exporter/app/command/exporter.go
	renamed:    maya-volume-exporter/main.go -> maya-exporter/main.go

1. Why is this change necessary ?
- Refer to issue openebs/openebs#1379

2. How does this change address the issue ?
- Rename maya-volume-exporter to maya-exporter wherever used in the code.
- Rename the same in GNumakefile, buildscripts and Dockerfile.

3. How to verify this change ?

- Run `maya-exporter --help`
- You should get the output similar to given below
```
maya-exporter can be used to monitor openebs volumes and pools.
It can be deployed alongside the openebs volume or pool containers as sidecars.

Usage:
   [flags]

Examples:
maya-exporter -a=http://localhost:8001 -c=:9500 -m=/metrics

Flags:
      --alsologtostderr                  log to standard error as well as files
  -c, --controller.addr string           Address of the Jiva volume controller. (default "http://localhost:9501")
  -h, --help                             help for this command
  -a, --listen.addr string               Address on which to expose metrics and web interface.) (default ":9500")
  -m, --listen.path string               Path under which to expose metrics. (default "/metrics")
      --log-flush-frequency duration     Maximum number of seconds between log flushes (default 5s)
      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
      --log_dir string                   If non-empty, write log files in this directory
      --logtostderr                      log to standard error instead of files (default true)
      --stderrthreshold severity         logs at or above this threshold go to stderr (default 2)
  -v, --v Level                          log level for V logs
      --vmodule moduleSpec               comma-separated list of pattern=N settings for file-filtered logging
```
4. What side effects does this change have ?
- none

5. Other details :
- Improvement: openebs/openebs#1379
