+++ buildscripts/exporter/Dockerfile
@@ -1,5 +1,5 @@
 FROM alpine:3.6
-ADD maya-volume-exporter /usr/local/bin
+ADD maya-exporter /usr/local/bin
 RUN apk add --no-cache \
      iproute2 \
      curl \
@@ -10,10 +10,10 @@ RUN apk add --no-cache \
 
 ARG BUILD_DATE
 LABEL org.label-schema.name="m-exporter"
-LABEL org.label-schema.description="OpenEBS Volume Exporter"
+LABEL org.label-schema.description="OpenEBS Maya Exporter"
 LABEL org.label-schema.url="http://www.openebs.io/"
 LABEL org.label-schema.vcs-url="https://github.com/openebs/maya"
 LABEL org.label-schema.schema-version="1.0"
 LABEL org.label-schema.build-date=$BUILD_DATE
-CMD maya-volume-exporter
+CMD maya-exporter
 EXPOSE 9500
